split package versions out of models
