+++ collectors/bzimport/tests/test_collectors.py
@@ -7,7 +7,8 @@ from django.utils import timezone
 from apps.bbsync.models import BugzillaComponent, BugzillaProduct
 from collectors.bzimport.collectors import BugzillaQuerier, MetadataCollector
 from collectors.bzimport.constants import BZ_DT_FMT
-from osidb.models import Affect, Flaw, FlawComment, Package, Tracker
+from osidb.dmodels.package_versions import Package
+from osidb.models import Affect, Flaw, FlawComment, Tracker
 from osidb.sync_manager import BZTrackerLinkManager
 from osidb.tests.factories import (
     AffectFactory,
