+++ osidb/tests/endpoints/flaws/test_unembargo.py
@@ -5,6 +5,7 @@ from django.utils import timezone
 from freezegun import freeze_time
 from rest_framework import status
 
+from osidb.dmodels.package_versions import Package
 from osidb.models import (
     Affect,
     AffectCVSS,
@@ -13,7 +14,6 @@ from osidb.models import (
     FlawComment,
     FlawCVSS,
     FlawReference,
-    Package,
     Tracker,
 )
 from osidb.tests.factories import (
